% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chris-datarelease.R
\name{chris_data_path}
\alias{chris_data_path}
\alias{list_data_modules}
\title{CHRIS data path}
\usage{
chris_data_path()

list_data_modules(path = chris_data_path())
}
\arguments{
\item{path}{\code{character(1)} specifying the directory where to look for CHRIS
data modules.}
}
\value{
See the individual function descriptions.
}
\description{
CHRIS data is organized in a hierarchical folder structure. The functions
documented here allow to define the base CHRIS data path and to list
available data modules.

The available functions are:
\itemize{
\item \code{chris_data_path}: returns a \code{character(1)} with the currently defined full
path to the CHRIS data.
\item \code{list_data_modules}: lists available data modules in the specified path.
}
}
\author{
Johannes Rainer
}
