% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chris-datarelease.R
\name{chrisDataPath}
\alias{chrisDataPath}
\alias{setChrisDataPath}
\alias{chrisDataReleases}
\alias{chrisDataRelease}
\alias{setChrisDataRelease}
\title{CHRIS data path and data releases}
\usage{
chrisDataPath()

setChrisDataPath(path = character())

chrisDataReleases(path = chrisDataPath())

chrisDataRelease()

setChrisDataRelease(release = chrisDataRelease())
}
\arguments{
\item{path}{\code{character(1)} defining the path where the CHRIS data can be
found.}

\item{release}{\code{character(1)} with the name of the CHRIS data release.}
}
\value{
See the individual function descriptions.
}
\description{
CHRIS data is organized in a hierarchical folder structure. The functions
documented here allow to define the base CHRIS data path or to list and
select available data releases.

The available functions are:
\itemize{
\item \code{chrisDataPath}: returns a \code{character(1)} with the currently defined full
path to the CHRIS data.
\item \code{setChrisDataPath}: sets/changes the path to the CHRIS data. The provided
\code{character(1)} needs to be the full path to the folder containing the
CHRIS data. Before setting the path, the function first evaluates if it
contains at least one valid data release folder.
\item \code{chrisDataReleases}: lists the data releases (folders) available in the
CHRIS data path.
\item \code{setChrisDataRelease}: defines the CHRIS release that should be used in the
current R session. Parameter \code{release} should be the name of the release
(which corresponds to the name of the folder in which the data modules for
the release are located).
\item \code{chrisDataRelease}: returns the currently set CHRIS release.
}
}
\author{
Johannes Rainer
}
