% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{export_tdf}
\alias{export_tdf}
\alias{labels_from_data}
\alias{mapping_from_data}
\alias{The}
\alias{`export_tdf`}
\alias{exports}
\alias{the}
\alias{provided}
\alias{data}
\alias{in}
\alias{TDF}
\alias{format.}
\alias{function}
\alias{first}
\alias{creates}
\alias{all}
\alias{required}
\alias{folders,}
\alias{checks}
\alias{input}
\alias{files}
\alias{and}
\alias{then}
\alias{format}
\alias{(see}
\alias{below}
\alias{for}
\alias{more}
\alias{information}
\alias{on}
\alias{this}
\alias{format).}
\alias{Automatic}
\alias{convertions}
\alias{performed}
\alias{by}
\alias{are:}
\alias{-}
\alias{Columns}
\alias{`data`}
\alias{that}
\alias{are}
\alias{of}
\alias{type}
\alias{`factor`}
\alias{correctly}
\alias{automatically}
\alias{converted}
\alias{to}
\alias{expected}
\alias{(i.e.}
\alias{their}
\alias{categories}
\alias{added}
\alias{`mapping`}
\alias{`data.frame`}
\alias{values}
\alias{replaced}
\alias{with}
\alias{indices).}
\alias{If}
\alias{not}
\alias{specified}
\alias{`labels`,}
\alias{columns}
\alias{`"min"`,}
\alias{`"max"`}
\alias{`labels`}
\alias{calculated}
\alias{`data`.}
\alias{Missing}
\alias{respective}
\alias{encoding}
\alias{`labels`.}
\alias{`labels_from_data`}
\alias{a}
\alias{*template*}
\alias{*labels*}
\alias{from}
\alias{retrieves}
\alias{various}
\alias{like}
\alias{adds}
\alias{corresponding}
\alias{`data.frame`.}
\alias{Other}
\alias{columns,}
\alias{such}
\alias{as}
\alias{`"description"`}
\alias{or}
\alias{`"unit"`}
\alias{need}
\alias{be}
\alias{filled}
\alias{out}
\alias{manually.}
\alias{`mapping_from_data`}
\alias{*mapping*}
\alias{categorical}
\alias{variables}
\alias{`factor`).}
\title{Export data in the Textual Dataset Format}
\usage{
export_tdf(
  name = character(),
  description = character(),
  version = character(),
  date = character(),
  path = ".",
  data = data.frame(),
  groups = data.frame(),
  grp_labels = data.frame(),
  labels = labels_from_data(data),
  mapping = mapping_from_data(data),
  na = -89
)

labels_from_data(data, na = -89)

mapping_from_data(data)
}
\arguments{
\item{name}{\strong{required} \code{character(1)} with the name of the data module.}

\item{description}{\code{character(1)} providing a description of the data.}

\item{version}{\strong{required} \code{character(1)} with the version of the data.}

\item{date}{\code{character(1)} providing the date of the data.}

\item{path}{\code{character(1)} with the base path where the folders and data
files should be created. Defaults to \code{path = "."}.}

\item{data}{\code{data.frame} with the data to export. Required column \code{"aid"} is
expected to contain the unique identifiers of the participants. All
additional columns are expected to contain the data of additional
variables.}

\item{groups}{\code{data.frame} with optional grouping of labels (variables) in
\code{data}. Expected columns are \code{"group"} and \code{"label"}. See the TDF
definition for details.}

\item{grp_labels}{\code{data.frame} with the names (descriptions) of the groups
defined in \code{groups}.}

\item{labels}{\code{data.frame} with \emph{annotations} to the variables (labels) in
\code{data}. See the TDF definition for details. Columns \code{"min"}, \code{"max"}
and \code{"missing"} will be filled by the \code{export_tdf} function if not
already provided. Defaults to \code{labels = labels_from_data()} hence creates
a \emph{labels} \code{data.frame} from the provided \code{data}.}

\item{mapping}{\code{data.frame} with the definition of the levels (categories)
of the categorical variables in \code{labels}. Expected columns are
\code{"label"}, \code{"code"} and \code{"value"}. The default is
\code{mapping = mapping_from_data(data)} and a \emph{mapping} \code{data.frame} is thus
generated by default from the provided \code{data}.}

\item{na}{the value to represent missing values in \code{data}.}
}
\value{
\code{export_tdf}: (invisibly) returns a \code{character(1)} with the path to
the folder where the data was stored. \code{labels_from_data} returns a
\emph{labels} \code{data.frame} based on the data in \code{data}.
}
\description{
Export data in the Textual Dataset Format
}
\section{Short information on the TDF}{


See the official
\href{https://wiki.gm.eurac.edu/index.php?title=Textual_Dataset_Format}{TDF}
definition for a complete description of the format.
\itemize{
\item \strong{data}: contains the data of the various variables. Columns are
variables, rows study participants. Column \code{"aid"} is mandatory and
contains the ID of the study participants.
\item \strong{labels}: provides information on the variables in \emph{data}. Columns are
\code{"label"} (the name of the column in \emph{data}), \code{"unit"} (unit of the
measured value), \code{"type"} (the data type), \code{"min"} (the minimal value),
\code{"max"} (the maximal value), \code{"missing"} (the value with which missing
values in \emph{data} are encoded) and \code{"description"} (a name/description of
the variable).
\item \strong{mapping}: contains the encoding of categorical variables (\code{factor}s) in
\emph{data}. Required columns are: \code{"label"} (the name of the column in \emph{data}),
\code{"code"} (the value of the category in \emph{data}) and \code{"value"} (the category,
i.e. the \code{level} of the \code{factor}).
\item \strong{groups}: allows to optionally group variables in \emph{data}. Expected
columns are \code{"group"} (the name of the group) and \code{"label"} (the name of
the column in \emph{data}).
\item \strong{grp_labels}: contains descriptions for the \emph{groups}. Expected columns
are \code{"group"} (the name of the group) and \code{"description"} (the
name/description of the group).
}
}

\examples{

## Exporting a test data set. Creating a *data* data.frame with data on
## 5 individuals.
d <- data.frame(
    aid = c("00101", "00102", "00103", "00104", "00105"),
    x0_sex = factor(c("Male", "Female", "Female", NA, "Male")),
    x0_age = c(45, 54, 33, 36, 66),
    x0_weight = c(78.5, 57.2, 55.2, 67.9, 84.2))

## Generate a *labels* data.frame from the data
l <- labels_from_data(d)
l

## Fill missing information to labels
l$unit <- c(NA, "Year", "kg")
l$description <- c("Sex", "Age", "Weight")

## Generate a *mapping* data.frame from data
m <- mapping_from_data(d)
m

## Create a simple grouping of all variables into a "general information"
## group
g <- data.frame(
    group = c("ginfo", "ginfo", "ginfo"),
    label = c("x0_sex", "x0_age", "x0_weight"))

## Define a description for the group
gl <- data.frame(group = "ginfo", description = "General information")

## Now export all data to a temporary folder
path <- tempdir()

## Export the data specifying the name of the module, the version and other
## information
export_tdf(name = "test_data", description = "Simple test data.",
    version = "1.0.0", date = date(), path = path, data = d,
    groups = g, grp_labels = gl, labels = l, mapping = m)
}
\author{
Johannes Rainer
}
