% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_aid}
\alias{format_aid}
\title{Format AIDs adding leading 0s}
\usage{
format_aid(x, length = 10)
}
\arguments{
\item{x}{\code{numeric} with the IDs to format.}

\item{length}{\code{integer(1)} defining the length (in characters) of the AID
format.}
}
\value{
\code{character} with the formated AID.
}
\description{
The \code{format_aid} function formats (numeric) AID IDs into character strings
of a predefined length (parameter \code{length}) adding leading \code{"0"}s to the
provided ID (\code{x}).
}
\examples{

format_aid(c(1, 2, 3))

## IDs provided as `character` are not formatted.
format_aid("hello")
}
\author{
Johannes Rainer
}
