% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R
\name{remove_participants}
\alias{remove_participants}
\title{Remove data for selected participants}
\usage{
remove_participants(x, aid, path = modulePath(x, base = TRUE))
}
\arguments{
\item{x}{\code{DataModule} from which data of a participant should be removed.}

\item{aid}{\code{character} with the IDs (label \code{"aid"}) of the participants for
whom data should be removed. Note that \code{\link[=format_aid]{format_aid()}} will be called
on this parameter to ensure the format of the AID matches.}

\item{path}{\code{character(1)} defining the path where the new version of the
data set should be stored into. Defaults to the base path of the
current module.}
}
\value{
\code{character(1)} with the path to the folder containing the new
version.
}
\description{
The \code{remove_participants} function creates a new version of an existing
data module removing any data for the selected individuals (parameter
\code{aid}). The function will:
\itemize{
\item bump the version of the module (incrementing the 4th position in the
\emph{x.y.z.a} version scheme).
\item create a new folder (name being the new version) in the directory \code{path}.
\item copy all files from the original module (version) to the new folder.
\item remove all entries (rows) in the \emph{data.txt} files with their \code{aid} label
matching \code{aid}.
\item validate the new data module.
}
}
\author{
Johannes Rainer
}
