% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{TDF}
\alias{TDF}
\alias{labels}
\title{Textual Dataset Format}
\description{
The \code{tidyfr} package provides convenience and utility functions to import
data stored in the new \emph{Textual Dateset Format} into R. These
functions ensure the data is properly formatted which include the correct
encoding categorical variables (\code{factor}s) or also missing values.

TDF data is structured in the following format:
\itemize{
\item \strong{data}: contains the data of the various variables. Columns are
variables, rows study participants. Column \code{"aid"} contains the ID of the
study participants.
\item \strong{labels}: provides information on the variables in \emph{data}. Columns are
\code{"label"} (the name of the column in \emph{data}), \code{"unit"} (unit of the
measured value), \code{"type"} (the data type), \code{"min"} (the minimal value),
\code{"max"} (the maximal value), \code{"missing"} (the value with which missing
values in \emph{data} are encoded) and \code{"description"} (a name/description of
the variable). Optional additional columns (annotations) might be
available depending on the data module.
\item \strong{groups}: provides optional grouping of variables in \emph{data}.
\item \strong{grp_labels}: contains descriptions for the \emph{groups}.
}

See the official
\href{https://wiki.gm.eurac.edu/index.php?title=Textual_Dataset_Format}{TDF}
definition for a complete description of the format.
}
\section{Accessing Data}{


The following functions can be used to import CHRIS data from one of the
available data modules.
\itemize{
\item \code{data}: retrieves the \emph{data}.
\item \code{groups}: retrieves the grouping information for variables.
\item \code{labels}: retrieves additional annotation/information for specific
variables of a module.
}

Which data is retrieved depends on the parameter \code{object}.
See \link{chris-SummarizedExperiment} if \code{object} is a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}}.
}

\author{
Johannes Rainer
}
