% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R
\name{valid_data_module}
\alias{valid_data_module}
\title{Data Module Validator}
\usage{
valid_data_module(path)
}
\arguments{
\item{path}{\code{character(1)} with the (full) path of the data module. The
specified directory should contain sub-folders with different versions
of the data module.}
}
\value{
The function returns (invisibly) \code{TRUE} if the data module is valid,
or throws an error.
}
\description{
The \code{valid_data_module} function takes a path to a TDFF data module as
input parameter and checks its validity. In detail, the function checks
the \emph{data} folder of each of the data module versions in \code{path} for
required data files and content.
}
\author{
Johannes Rainer
}
