% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment.R
\name{chris-SummarizedExperiment}
\alias{chris-SummarizedExperiment}
\alias{labels,SummarizedExperiment-method}
\alias{groups,SummarizedExperiment-method}
\alias{groups.SummarizedExperiment}
\title{Extracting TDF-compliant data from a SummarizedExperiment}
\usage{
\method{groups}{SummarizedExperiment}(x, assayNames. = assayNames(x), labelPrefix = "x0xx", ...)

\S4method{labels}{SummarizedExperiment}(object, assayNames. = assayNames(object), labelPrefix = "x0xx")
}
\arguments{
\item{x}{a \code{SummarizedExperiment} object}

\item{assayNames.}{\code{character} defining the names of the \emph{assays} in
\code{object} from which data should be extracted.}

\item{labelPrefix}{\code{character(1)} defining the prefix for the variable IDs
(labels) of the object.}

\item{...}{Further arguments passed to downstream \code{groups} method}

\item{object}{A \code{SummarizedExperiment} object.}
}
\value{
a \code{data.frame} with the data.
}
\description{
The \code{\link[=SummarizedExperiment]{SummarizedExperiment()}} class is a container for data from large scale
assays for biological experiments. In contrast to TDF data, samples are
organized in columns of a \code{SummarizedExperiment} and measurements in rows.
The \code{data}, \code{labels} and \code{groups} methods allow to extract information from
such objects in a Textual Dataset File (TDF)-compliant format
(structure):
\itemize{
\item columns are variables, rows individuals (samples).
\item variable IDs (labels) follow a standard format (e.g. \code{"x0pt001"}).
}

The available methods are:
\itemize{
\item \code{data}: export the (quantitative) assay data from an
\code{SummarizedExperiment} as a \code{data.frame} with columns representing
variables and rows samples (study participants). Parameter \code{assayNames.}
allows to specify which of the assays from the \code{SummarizedExperiment}
should be extracted. Variables (rows) in the \code{SummarizedExperiment} get
assigned a new variable ID (called \emph{label}), which consists of the
\code{labelPrefix} followed by an integer representing the index of the
variable in the \code{SummarizedExperiment} (i.e. the row number of the
variable in the \code{SummarizedExperiment}). A letter is appended to IDs for
variables from assays different than the first one. Thus, \code{"x0xx001"}
corresponds to the first row in the first assay, while \code{"x0xx001a"}
represents the first row in the second assay. The \code{colnames} of the
\code{SummarizedExperiment} are used as sample identifiers and are returned in
column \code{"aid"} of the result \code{data.frame}.
\item \code{groups}: retrieves a \code{data.frame} that specifies the grouping of
variables returned by \code{data} from a \code{SummarizedExperiment}. Columns
(variables) containing data from the same \code{assay} of the
\code{SummarizedExperiment} are grouped into the same group.
\item \code{labels}: extracts label annotations for the data extracted with
\code{data} from a \code{SummarizedExperiment}. The returned \code{data.frame} is in the
\emph{labels} format of the TDF but contains additional columns with the
available annotations from the \code{SummarizedExperiment}'s \code{\link[=rowData]{rowData()}}. The
\code{rownames} of the \code{SummarizedExperiment} are returned in columns
\code{"description"}.
}
}
\examples{

## Create a simple SummarizedExperiment with some random data as one assay
## and a second assay with all values multiplied with 2. For a
## SummarizedExperiment columns represent samples and rows measurements
## (variables).
mat <- matrix(rnorm(60), ncol = 10, nrow = 6)

## SummarizedExperiments allow to store also column and row annotations along
## with the data. We thus define below a data.frame with some annotations
## for the variables.
rowd <- data.frame(analyte_id = paste0("id", 1:6), analyte_name = letters[1:6])
rownames(rowd) <- rowd$analyte_id

library(SummarizedExperiment)
se <- SummarizedExperiment(
    assay = list(values = mat, double = 2 * mat),
    rowData = rowd)
se

## What assays are available?
assayNames(se)

## Get a data.frame with all variables
data(se)

## Get the label information
labels(se)

## Get the variable grouping
groups(se)
}
\author{
Johannes Rainer
}
