% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R
\name{DataModule}
\alias{DataModule}
\alias{DataModule-class}
\alias{show,DataModule-method}
\alias{data_module}
\alias{moduleName}
\alias{modulePath}
\alias{moduleVersion}
\alias{moduleDescription}
\alias{moduleDate}
\title{Data Modules}
\usage{
data_module(name = character(), version = character(), path = data_path())

moduleName(object)

modulePath(object)

moduleVersion(object)

moduleDescription(object)

moduleDate(object)
}
\arguments{
\item{name}{For \code{data_module}: \code{character(1)} defining the name of the
module to load.}

\item{version}{For \code{data_module}: \code{character(1)} defining the version of the
module to load.}

\item{path}{For \code{data_module}: \code{character(1)} defining the path where data
modules are stored.}

\item{object}{A \code{DataModule} object.}
}
\value{
See the individual function description.
}
\description{
A Data Module provides the data of one specific module, which can be
the interview, clinical blood parameters or the metabolomics or proteomics
data sets. The actual data from a module is stored in the Textual Dataset
Format (TDF - see \link{TDF} for more details).

The \code{tidyfr} package represents a data module with the \code{DataModule} object
which provides all necessary functionality to import data of a module and to
format it properly for R.
}
\section{Loading a module}{


Available data modules in a certain path can be listed using the
\code{\link[=list_data_modules]{list_data_modules()}} function.
\itemize{
\item \code{data_module}: load a specific data module. The name and version of the
data module to load needs to be specified with parameters \code{name} and
\code{version} respectively. Parameter \code{path} can be used to set the base path
where the data module can be found. The function returns an instance of
\code{DataModule}.
}
}

\section{Accessing properties and data from a module}{

\itemize{
\item \code{moduleName}: returns the name of a module.
\item \code{modulePath}: returns the (full) file path to the data module.
\item \code{moduleVersion}: returns the version of the data module.
\item \code{moduleDescription}: returns the description of the module.
\item \code{moduleDate}: returns the date of the module.
}
}

\examples{

## List available test data modules provided by the tidyfr package
pth <- system.file("txt", package = "tidyfr")
list_data_modules(pth)

## Load one data module
mdl <- data_module(name = "db_example2", version = "1.0.1", path = pth)
mdl

## Get the name, description and version of the module
moduleName(mdl)
moduleDescription(mdl)
moduleVersion(mdl)
}
\author{
Johannes Rainer
}
