% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R
\name{DataModule}
\alias{DataModule}
\alias{DataModule-class}
\alias{moduleName}
\alias{modulePath}
\alias{moduleVersion}
\alias{moduleDescription}
\alias{moduleDate}
\title{Data Modules}
\usage{
moduleName(object)

modulePath(object)

moduleVersion(object)

moduleDescription(object)

moduleDate(object)
}
\arguments{
\item{object}{A \code{DataModule} object.}
}
\value{
See the individual function description.
}
\description{
A Data Module provides the data of one specific module, which can be
the interview, clinical blood parameters or the metabolomics or proteomics
data sets. The actual data from a module is stored in the Textual Dataset
Format (TDF - see \link{TDF} for more details).

The \code{tidyfr} package represents a data module with the \code{DataModule} object
which provides all necessary functionality to import data of a module and to
format it properly for R.
}
\section{Loading a module}{


Available data modules in a certain path can be listed using the
\code{\link[=list_data_modules]{list_data_modules()}} function.
}

\section{Accessing properties and data from a module}{

\itemize{
\item \code{moduleName}: returns the name of a module.
\item \code{modulePath}: returns the (full) file path to the data module.
\item \code{moduleVersion}: returns the version of the data module.
\item \code{moduleDescription}: returns the description of the module.
\item \code{moduleDate}: returns the date of the module.
}
}

\author{
Johannes Rainer
}
